refactor(web/submissions): split lib/submissions.ts into a directory
Moves submissions.ts (473 LOC after the moderator wiring) under
the loc-guardian 350-LOC ceiling by extracting six files with
single responsibilities:

  schema.ts  — input/output types and zod validators (84 LOC)
  state.ts   — loadAuthorContext, karma gate, dedup query (82)
  create.ts  — createSubmission with the moderator pipeline (148)
  edit.ts    — updateSubmissionAsAuthor (155)
  delete.ts  — deleteSubmissionAsAuthor (34)
  index.ts   — barrel of the public surface (34)

The barrel keeps `@/lib/submissions` callers untouched — all seven
existing importers (REST routes, MCP tools, web action, DTO
mapper, input parsers) resolve through index.ts unchanged.

Helpers that were previously module-private (loadAuthorContext,
findRecentDuplicate, the karma-gate constants) stay private to
the directory — only schema/types and the three lifecycle verbs
re-export.

No behavioral change. Pre-existing 378-LOC overshoot was inherited
from before this branch; the moderator integration pushed it to
473, which is what triggered the split.
